+++ app/Extensions/Plugin/BasePluginServiceProvider.php
@@ -73,7 +73,7 @@ abstract class BasePluginServiceProvider extends ServiceProvider
     {
         $migrationsPath = $this->pluginPath.'/database/migrations';
 
-        if (File::exists($migrationsPath) && $this->app->runningInConsole()) {
+        if (File::exists($migrationsPath)) {
             $this->loadMigrationsFrom($migrationsPath);
         }
     }